+++ src/lisp.rs
@@ -7,27 +7,44 @@
 
 use std::collections::HashMap;
 use std::fmt::Write;
+use std::sync::{Arc, RwLock};
 use std::{error, fmt, iter};
 
+use crate::strmap::{MappedStr, StrMap};
 use crate::utils::safe_string;
 
 #[derive(Clone)]
 pub enum AsylType {
 	Symbol, String, Float, Int, Bool, Type, List, Fn
 }
+impl fmt::Display for AsylType {
+	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+		f.write_str(match self {
+			AsylType::Symbol => "'symbol",
+			AsylType::String => "'string",
+			AsylType::Float  => "'float",
+			AsylType::Int    => "'int",
+			AsylType::Bool   => "'bool",
+			AsylType::Type   => "'type",
+			AsylType::List   => "'list",
+			AsylType::Fn     => "'fn",
+		})
+	}
+}
+#[derive(Clone)]
 pub struct AsylLambda {
 	// todo
 	// i dont think these need to be rc'd but suuure
 	// i can most likely get away with boxing these
 	// i'll try that when i get home
-	args: Rc<Vec<String>>,
-	body: Rc<Spanned<AsylExpr>>,
+	args: Vec<MappedStr>,
+	body: AsylExpr,
+	env: Arc<RwLock<AsylEnv>>,
 }
-// string interning could be nice here, probably implemented as an Rc<String>?
 #[derive(Clone)]
-pub enum AsylExpr {
-	Symbol(String),
-	String(String),
+pub enum AsylExprValue {
+	Symbol(MappedStr),
+	String(MappedStr),
 	Float(f64),
 	Int(i64),
 	Bool(bool),
@@ -40,45 +57,46 @@ pub enum AsylExpr {
 	//     a linked list of lists
 	// - VList
 	//   - probably a slightly better implementation of that snap hybrid list
-	List(Vec<Spanned<AsylExpr>>),
-	ExtFn(fn(AsylSpan, &mut AsylEnv, &[Spanned<AsylExpr>]) -> Result<Spanned<AsylExpr>, AsylError>),
-	Lambda(AsylLambda),
+	List(Vec<AsylExpr>),
+	ExtFn(fn(AsylSpan, &Arc<RwLock<AsylEnv>>, &[AsylExpr], &mut StrMap) -> Result<AsylExpr, AsylError>),
+	Lambda(Box<AsylLambda>),
 }
 
-impl AsylExpr {
+impl AsylExprValue {
 	fn get_type(&self) -> AsylType {
 		match self {
-			AsylExpr::Symbol(_) => AsylType::Symbol,
-			AsylExpr::String(_) => AsylType::String,
-			AsylExpr::Float(_)  => AsylType::Float,
-			AsylExpr::Int(_)    => AsylType::Int,
-			AsylExpr::Bool(_)   => AsylType::Bool,
-			AsylExpr::Type(_)   => AsylType::Type,
-			AsylExpr::List(_)   => AsylType::List,
-			AsylExpr::ExtFn(_)  => AsylType::Fn,
-			AsylExpr::Lambda(_) => AsylType::Fn,
+			AsylExprValue::Symbol(_) => AsylType::Symbol,
+			AsylExprValue::String(_) => AsylType::String,
+			AsylExprValue::Float(_)  => AsylType::Float,
+			AsylExprValue::Int(_)    => AsylType::Int,
+			AsylExprValue::Bool(_)   => AsylType::Bool,
+			AsylExprValue::Type(_)   => AsylType::Type,
+			AsylExprValue::List(_)   => AsylType::List,
+			AsylExprValue::ExtFn(_)  => AsylType::Fn,
+			AsylExprValue::Lambda(_) => AsylType::Fn,
 		}
 	}
 }
 
-// char, line, col
-#[derive(Debug, Copy, Clone)]
-pub struct AsylPos(usize, usize, usize);
+// char, line, col, name, content
+#[derive(Debug, Clone)]
+pub struct AsylPos(usize, usize, usize, MappedStr, MappedStr);
 // start, len
-#[derive(Debug, Copy, Clone)]
+#[derive(Debug, Clone)]
 pub struct AsylInSpan(AsylPos, usize);
 type AsylSpan = Option<AsylInSpan>;
 
 impl AsylInSpan {
-	pub fn print(&self, source_name: &str, source: &str) -> Result<String, fmt::Error> {
+	pub fn print(&self) -> Result<String, fmt::Error> {
+		let (source_name, source) = (&self.0.3, &self.0.4);
 		let mut out = String::from("```");
 		let mut cols_until_start = self.0.2 - 1;
 		let mut line = self.0.1;
 		let mut current_col = 1;
-		writeln!(out, "{}:{}", source_name, self.0.1)?;
+		writeln!(out, "{}:{}", source_name.get_ref(), self.0.1)?;
 		write!(out, "{:>6} | ", line)?;
 		let mut count = self.1;
-		for (_, ch) in source.chars().chain(iter::once('\n')).skip(self.0.0 + 1 - self.0.2).enumerate() {
+		for (_, ch) in source.get_ref().chars().chain(iter::once('\n')).skip(self.0.0 + 1 - self.0.2).enumerate() {
 			match ch {
 				'\n' => {
 					line += 1;
@@ -110,16 +128,17 @@ impl AsylInSpan {
 
 impl fmt::Display for AsylExpr {
 	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		f.write_str(&match self {
-			AsylExpr::Symbol(v) => v.clone(),
-			AsylExpr::String(v) => format!("{:?}", v),
-			AsylExpr::Float(v) => v.to_string(),
-			AsylExpr::Int(v) => v.to_string(),
+		f.write_str(&match &self.0 {
+			AsylExprValue::Symbol(v) => v.to_string(),
+			AsylExprValue::String(v) => format!("{:?}", v.to_string()),
+			AsylExprValue::Float(v) => v.to_string(),
+			AsylExprValue::Int(v) => v.to_string(),
 			// technically 't and 'f are just variables but i don't care
-			AsylExpr::Bool(v) => if *v {"'t"} else {"'f"}.to_string(),
-			AsylExpr::List(v) => format!("({})", v.iter().map(|v| v.0.to_string()).collect::<Vec<_>>().join(" ")),
-			AsylExpr::ExtFn(_) => "<proc>".to_string(),
-			AsylExpr::Lambda(_) => "<proc>".to_string(),
+			AsylExprValue::Bool(v) => if *v {"'t"} else {"'f"}.to_string(),
+			AsylExprValue::List(v) => format!("({})", v.iter().map(|v| v.to_string()).collect::<Vec<_>>().join(" ")),
+			AsylExprValue::ExtFn(_) => "<proc>".to_string(),
+			AsylExprValue::Lambda(_) => "<proc>".to_string(),
+			AsylExprValue::Type(v) => v.to_string(),
 		})
 	}
 }
@@ -132,52 +151,44 @@ impl fmt::Debug for AsylExpr {
 
 // todo: replace the strings with more specific details (a range, asyltype, etc)
 #[derive(Debug)]
-pub enum AsylError {
+pub enum AsylErrorType {
 	UnexpectedEOF,
-	InvalidEscape(AsylSpan, String),
-	UnexpectedCloseParen(AsylSpan),
-	InvalidNumber(AsylSpan, String),
-	TypeMismatch(AsylSpan, String),
-	ArgMismatch(AsylSpan, String),
-	NotDefined(AsylSpan, String),
-	InvalidCall(AsylSpan),
+	InvalidEscape(String),
+	UnexpectedCloseParen,
+	InvalidNumber(String),
+	TypeMismatch(String),
+	ArgMismatch(String),
+	NotDefined(String),
+	InvalidCall,
+	Internal,
 	Shit,
 }
 
+#[derive(Debug)]
+pub struct AsylError(AsylErrorType, AsylSpan);
+
 impl fmt::Display for AsylError {
 	fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-		f.write_str(&match self {
-			AsylError::UnexpectedEOF => "Unexpected end of input".to_string(),
-			AsylError::InvalidEscape(_, text) => format!("Invalid escape sequence \\`{}\\`", safe_string(text)),
-			AsylError::UnexpectedCloseParen(_) => "Unexpected `)`".to_string(),
-			AsylError::InvalidNumber(_, text) => format!("Invalid number \\`{}\\`", safe_string(text)),
-			AsylError::ArgMismatch(_, text) => format!("Argument mismatch, expected {}", text),
-			AsylError::NotDefined(_, text) => format!("\\`{}\\` isn't defined", safe_string(text)),
-			AsylError::InvalidCall(_) => "Invalid call".to_string(),
-			AsylError::TypeMismatch(_, text) => format!("Type mismatch, expected {}", text),
-			AsylError::Shit() => "not implemented".to_string(),
+		f.write_str(&match &self.0 {
+			AsylErrorType::UnexpectedEOF => "Unexpected end of input".to_string(),
+			AsylErrorType::InvalidEscape(text) => format!("Invalid escape sequence \\`{}\\`", safe_string(text)),
+			AsylErrorType::UnexpectedCloseParen => "Unexpected `)`".to_string(),
+			AsylErrorType::InvalidNumber(text) => format!("Invalid number \\`{}\\`", safe_string(text)),
+			AsylErrorType::ArgMismatch(text) => format!("Argument mismatch, expected {}", text),
+			AsylErrorType::NotDefined(text) => format!("\\`{}\\` isn't defined", safe_string(text)),
+			AsylErrorType::InvalidCall => "Invalid call".to_string(),
+			AsylErrorType::TypeMismatch(text) => format!("Type mismatch, expected {}", text),
+			AsylErrorType::Internal => "Internal error".to_string(),
+			AsylErrorType::Shit => "Not implemented".to_string(),
 		})
 	}
 }
 
 impl AsylError {
-	pub fn span(&self) -> AsylSpan {
-		*match self {
-			AsylError::UnexpectedEOF => &None,
-			AsylError::InvalidEscape(s, _) => s,
-			AsylError::UnexpectedCloseParen(s) => s,
-			AsylError::InvalidNumber(s, _) => s,
-			AsylError::TypeMismatch(s, _) => s,
-			AsylError::ArgMismatch(s, _) => s,
-			AsylError::NotDefined(s, _) => s,
-			AsylError::InvalidCall(s) => s,
-			AsylError::Shit => None,
-		}
-	}
-	pub fn print_span(&self, source_name: &str, source: &str) -> String {
-		match self.span() {
-			Some(v) => v.print(source_name, source).unwrap_or("no trace".to_string()),
-			None => "no trace".to_string()
+	pub fn print(&self) -> String {
+		match &self.1 {
+			Some(v) => v.print().unwrap_or("no trace :(".to_string()),
+			None => "no trace :(".to_string()
 		}
 	}
 }
@@ -185,17 +196,17 @@ impl AsylError {
 impl error::Error for AsylError {}
 
 type AsylResult<T> = Result<T, AsylError>;
-#[derive(Clone)]
+#[derive(Debug, Clone)]
 pub enum AsylEnvEntry {
 	// once you've just defined something
-	Unevaluated(Spanned<AsylExpr>),
+	Unevaluated(AsylExpr),
 	// if you use a value
-	Evaluated(Spanned<AsylExpr>),
+	Evaluated(AsylExpr),
 }
-#[derive(Clone)]
-pub struct AsylEnv<'a> {
-	data: HashMap<String, AsylEnvEntry>,
-	parent: Option<&'a mut AsylEnv<'a>>,
+#[derive(Debug, Clone)]
+pub struct AsylEnv {
+	data: HashMap<MappedStr, AsylEnvEntry>,
+	parent: Option<Arc<RwLock<AsylEnv>>>,
 }
 
 #[derive(Debug)]
@@ -204,16 +215,16 @@ enum UniformNumberList {
 	Int(Vec<i64>),
 }
 
-fn to_uniform_number_list(list: &[Spanned<AsylExpr>]) -> AsylResult<UniformNumberList> {
+fn to_uniform_number_list(list: &[AsylExpr]) -> AsylResult<UniformNumberList> {
 	#[derive(PartialEq, Eq, PartialOrd, Ord)]
 	enum NumType {
 		Float, Int
 	}
-	fn num_type(v: &Spanned<AsylExpr>) -> AsylResult<NumType> {
+	fn num_type(v: &AsylExpr) -> AsylResult<NumType> {
 		match v.0 {
-			AsylExpr::Float(_) => Ok(NumType::Float),
-			AsylExpr::Int(_)   => Ok(NumType::Int),
-			_ => Err(AsylError::TypeMismatch(v.1, "'float or 'int".to_string()))
+			AsylExprValue::Float(_) => Ok(NumType::Float),
+			AsylExprValue::Int(_)   => Ok(NumType::Int),
+			_ => Err(AsylError(AsylErrorType::TypeMismatch("'float or 'int".to_string()), v.1.clone()))
 		}
 	}
 	let mut best = NumType::Int;
@@ -228,8 +239,8 @@ fn to_uniform_number_list(list: &[Spanned<AsylExpr>]) -> AsylResult<UniformNumbe
 			let mut res = vec![];
 			for i in list {
 				res.push(match i.0 {
-					AsylExpr::Int($a)   => $int,
-					$(AsylExpr::Float($a) => $float,
+					AsylExprValue::Int($a)   => $int,
+					$(AsylExprValue::Float($a) => $float,
 					)? _ => unreachable!(),
 				});
 			}
@@ -256,171 +267,178 @@ macro_rules! tonicity_internal {
 // this is clever, so i'm stealing it :)
 macro_rules! ensure_tonicity {
 	($check_fn:expr) => {
-		|this, env, args| {
-			let args = eval_args(args, env)?;
+		|this, env, args, map| {
+			let args = eval_args(args, env, map)?;
 			asal!(this, args, 1.., "at least one argument");
 			tonicity_internal!(f64, f_f64, $check_fn);
 			tonicity_internal!(i64, f_i64, $check_fn);
-			Ok(Spanned(AsylExpr::Bool(match to_uniform_number_list(&args)? {
+			Ok(AsylExpr(AsylExprValue::Bool(match to_uniform_number_list(&args)? {
 				UniformNumberList::Float(v) => f_f64(&v[0], &v[1..]),
 				UniformNumberList::Int(v)   => f_i64(&v[0], &v[1..]),
-			}), this))
+			}), this, None))
 		}
 	};
 }
 
 macro_rules! ast {
 	($this:expr, $val:expr, $ty:tt, $ty_str:expr) => {
-		if let AsylExpr::$ty(v) = $val.0 {
+		if let AsylExprValue::$ty(v) = &$val.0 {
 			Ok(v)
 		} else {
-			Err(AsylError::TypeMismatch($this, $ty_str.to_string()))
+			Err(AsylError(AsylErrorType::TypeMismatch($ty_str.to_string()), $this))
 		}
 	}
 }
 
 macro_rules! asal {
 	($this:expr, $args:expr, $range:expr, $msg:expr) => {
-		if !$range.contains($args.len()) {
-			return Err(AsylError::ArgMismatch($this, $msg))
+		if !$range.contains(&$args.len()) {
+			return Err(AsylError(AsylErrorType::ArgMismatch($msg.to_string()), $this))
 		}
 	}
 }
 
 macro_rules! envi {
-	($env:expr, $name:expr, $value:expr) => {
-		$env.insert($name.to_string(), Spanned($value, None))
+	($map:expr, $env:expr, $name:expr, $value:expr) => {
+		$env.insert(MappedStr::new($name, $map), AsylEnvEntry::Evaluated(AsylExpr($value, None, None)))
 	}
 }
 
-pub fn default_env() -> AsylEnv {
+pub fn default_env(map: &mut StrMap) -> AsylEnv {
 	let mut data = HashMap::new();
 	// maths
-	envi!(data, "+", AsylExpr::ExtFn(|this, env, args| {
-		let args = eval_args(args, env)?;
+	envi!(map, data, "+", AsylExprValue::ExtFn(|this, env, args, map| {
+		let args = eval_args(args, env, map)?;
 		asal!(this, args, 1.., "at least one argument");
-		Ok(Spanned(match to_uniform_number_list(&args)? {
-			UniformNumberList::Float(v) => AsylExpr::Float(v.iter().fold(0.0, |a, b| a + b)),
-			UniformNumberList::Int(v)   => AsylExpr::Int(  v.iter().fold(0, |a, b| a + b)),
-		}, this))
+		Ok(AsylExpr(match to_uniform_number_list(&args)? {
+			UniformNumberList::Float(v) => AsylExprValue::Float(v.iter().fold(0.0, |a, b| a + b)),
+			UniformNumberList::Int(v)   => AsylExprValue::Int(  v.iter().fold(0, |a, b| a + b)),
+		}, this, None))
 	}));
-	envi!(data, "*", AsylExpr::ExtFn(|this, env, args| {
-		let args = eval_args(args, env)?;
+	envi!(map, data, "*", AsylExprValue::ExtFn(|this, env, args, map| {
+		let args = eval_args(args, env, map)?;
 		asal!(this, args, 1.., "at least one argument");
-		Ok(Spanned(match to_uniform_number_list(&args)? {
-			UniformNumberList::Float(v) => AsylExpr::Float(v.iter().fold(1.0, |a, b| a * b)),
-			UniformNumberList::Int(v)   => AsylExpr::Int(  v.iter().fold(1, |a, b| a * b)),
-		}, this))
+		Ok(AsylExpr(match to_uniform_number_list(&args)? {
+			UniformNumberList::Float(v) => AsylExprValue::Float(v.iter().fold(1.0, |a, b| a * b)),
+			UniformNumberList::Int(v)   => AsylExprValue::Int(  v.iter().fold(1, |a, b| a * b)),
+		}, this, None))
 	}));
-	envi!(data, "-", AsylExpr::ExtFn(|this, env, args| {
-		let args = eval_args(args, env)?;
+	envi!(map, data, "-", AsylExprValue::ExtFn(|this, env, args, map| {
+		let args = eval_args(args, env, map)?;
 		asal!(this, args, 2.., "at least two arguments");
-		Ok(Spanned(match to_uniform_number_list(&args)? {
-			UniformNumberList::Float(v) => AsylExpr::Float(v[0] - v[1..].iter().fold(0.0, |a, b| a + b)),
-			UniformNumberList::Int(v)   => AsylExpr::Int(  v[0] - v[1..].iter().fold(0, |a, b| a + b)),
-		}, this))
+		Ok(AsylExpr(match to_uniform_number_list(&args)? {
+			UniformNumberList::Float(v) => AsylExprValue::Float(v[0] - v[1..].iter().fold(0.0, |a, b| a + b)),
+			UniformNumberList::Int(v)   => AsylExprValue::Int(  v[0] - v[1..].iter().fold(0, |a, b| a + b)),
+		}, this, None))
 	}));
 	// ' isn't quoting in this lang and # is used by discord
-	envi!(data, "'f", AsylExpr::Bool(false));
-	envi!(data, "'t", AsylExpr::Bool(true));
-	envi!(data, "'n", AsylExpr::List(vec![]));
+	envi!(map, data, "'f", AsylExprValue::Bool(false));
+	envi!(map, data, "'t", AsylExprValue::Bool(true));
+	envi!(map, data, "'n", AsylExprValue::List(vec![]));
 	// types
-	envi!(data, "'symbol", AsylExpr::Type(AsylType::Symbol));
-	envi!(data, "'string", AsylExpr::Type(AsylType::String));
-	envi!(data, "'float",  AsylExpr::Type(AsylType::Float));
-	envi!(data, "'int",    AsylExpr::Type(AsylType::Int));
-	envi!(data, "'bool",   AsylExpr::Type(AsylType::Bool));
-	envi!(data, "'type",   AsylExpr::Type(AsylType::Type));
-	envi!(data, "'list",   AsylExpr::Type(AsylType::List));
-	envi!(data, "'fn",     AsylExpr::Type(AsylType::Fn));
+	envi!(map, data, "'symbol", AsylExprValue::Type(AsylType::Symbol));
+	envi!(map, data, "'string", AsylExprValue::Type(AsylType::String));
+	envi!(map, data, "'float",  AsylExprValue::Type(AsylType::Float));
+	envi!(map, data, "'int",    AsylExprValue::Type(AsylType::Int));
+	envi!(map, data, "'bool",   AsylExprValue::Type(AsylType::Bool));
+	envi!(map, data, "'type",   AsylExprValue::Type(AsylType::Type));
+	envi!(map, data, "'list",   AsylExprValue::Type(AsylType::List));
+	envi!(map, data, "'fn",     AsylExprValue::Type(AsylType::Fn));
 	// comparisons
-	envi!(data, ">",  AsylExpr::ExtFn(ensure_tonicity!(|a, b| a > b)));
-	envi!(data, "=",  AsylExpr::ExtFn(ensure_tonicity!(|a, b| a == b)));
-	envi!(data, ">=", AsylExpr::ExtFn(ensure_tonicity!(|a, b| a >= b)));
-	envi!(data, "<",  AsylExpr::ExtFn(ensure_tonicity!(|a, b| a < b)));
-	envi!(data, "!=", AsylExpr::ExtFn(ensure_tonicity!(|a, b| a != b)));
-	envi!(data, "<=", AsylExpr::ExtFn(ensure_tonicity!(|a, b| a <= b)));
+	envi!(map, data, ">",  AsylExprValue::ExtFn(ensure_tonicity!(|a, b| a > b)));
+	envi!(map, data, "=",  AsylExprValue::ExtFn(ensure_tonicity!(|a, b| a == b)));
+	envi!(map, data, ">=", AsylExprValue::ExtFn(ensure_tonicity!(|a, b| a >= b)));
+	envi!(map, data, "<",  AsylExprValue::ExtFn(ensure_tonicity!(|a, b| a < b)));
+	envi!(map, data, "!=", AsylExprValue::ExtFn(ensure_tonicity!(|a, b| a != b)));
+	envi!(map, data, "<=", AsylExprValue::ExtFn(ensure_tonicity!(|a, b| a <= b)));
 	// logical joiners
-	envi!(data, "&&", AsylExpr::ExtFn(|this, env, args| {
+	envi!(map, data, "&&", AsylExprValue::ExtFn(|this, env, args, map| {
 		for arg in args {
-			let res = eval(arg, env)?;
-			if ast!(this, res, Bool, "'bool")? == false {
-				return Ok(Spanned(AsylExpr::Bool(false), this))
+			let res = eval(arg, env, map)?;
+			if *ast!(this.clone(), res, Bool, "'bool")? == false {
+				return Ok(AsylExpr(AsylExprValue::Bool(false), this, None))
 			}
 		}
-		Ok(Spanned(AsylExpr::Bool(true), this))
+		Ok(AsylExpr(AsylExprValue::Bool(true), this, None))
 	}));
-	envi!(data, "||", AsylExpr::ExtFn(|this, env, args| {
+	envi!(map, data, "||", AsylExprValue::ExtFn(|this, env, args, map| {
 		for arg in args {
-			let res = eval(arg, env)?;
-			if ast!(this, res, Bool, "'bool")? == true {
-				return Ok(Spanned(AsylExpr::Bool(true), this))
+			let res = eval(arg, env, map)?;
+			if *ast!(this.clone(), res, Bool, "'bool")? == true {
+				return Ok(AsylExpr(AsylExprValue::Bool(true), this, None))
 			}
 		}
-		Ok(Spanned(AsylExpr::Bool(false), this))
+		Ok(AsylExpr(AsylExprValue::Bool(false), this, None))
 	}));
-	envi!(data, "^^", AsylExpr::ExtFn(|this, env, args| {
-		let args = eval_args(args, env)?;
-		Ok(Spanned(AsylExpr::Bool(
+	envi!(map, data, "^^", AsylExprValue::ExtFn(|this, env, args, map| {
+		let args = eval_args(args, env, map)?;
+		Ok(AsylExpr(AsylExprValue::Bool(
 			args
 			.iter()
-			.map(|v| ast!(this, v, Bool, "'bool"))
+			.map(|v| ast!(this.clone(), v, Bool, "'bool"))
 			.collect::<AsylResult<Vec<_>>>()?
 			.iter()
-			.fold(false, |a, b| if *b {!a} else {a})
-		), this))
+			.fold(false, |a, b| if **b {!a} else {a})
+		), this, None))
 	}));
-	envi!(data, "!", AsylExpr::ExtFn(|this, env, args| {
-		let args = eval_args(args, env)?;
+	envi!(map, data, "!", AsylExprValue::ExtFn(|this, env, args, map| {
+		let args = eval_args(args, env, map)?;
 		asal!(this, args, 1..=1, "one argument");
-		Ok(Spanned(AsylExpr::Bool(!ast!(this, args[0], Bool, "'bool")?), this))
+		Ok(AsylExpr(AsylExprValue::Bool(!*ast!(this.clone(), args[0], Bool, "'bool")?), this, None))
 	}));
 	// typeof operator
-	envi!(data, "'", AsylExpr::ExtFn(|this, env, args| {
+	envi!(map, data, "'", AsylExprValue::ExtFn(|this, env, args, map| {
 		asal!(this, args, 1..=1, "one argument");
-		return Ok(AsylExpr::Type(args[0].get_type()))
+		return Ok(AsylExpr(AsylExprValue::Type(eval(&args[0], env, map)?.0.get_type()), this, None))
 	}));
-	envi!(data, "if", AsylExpr::ExtFn(|this, env, args| {
+	envi!(map, data, "if", AsylExprValue::ExtFn(|this, env, args, map| {
 		asal!(this, args, 3..=3, "three arguments");
-		if ast!(this, eval(&args[0], env)?, Bool, "'bool")? {
-			eval(&args[1], env)
+		if *ast!(this, eval(&args[0], env, map)?, Bool, "'bool")? {
+			eval(&args[1], env, map)
 		} else {
-			eval(&args[2], env)
+			eval(&args[2], env, map)
 		}
 	}));
 	// i like def and fn over define and lambda, so i'm using those
-	envi!(data, "def", AsylExpr::ExtFn(|this, env, args| {
+	envi!(map, data, "def", AsylExprValue::ExtFn(|this, env, args, _map| {
 		asal!(this, args, 2..=2, "two arguments");
-		let name = ast!(this, args[0], Symbol, "'symbol")?;
-		env.data.insert(name, AsylEnvEntry::Unevaluated(args[1].clone()));
+		let name = ast!(this.clone(), args[0], Symbol, "'symbol")?;
+		{
+			let mut lock = env.write().map_err(|_| AsylError(AsylErrorType::Internal, this))?;
+			lock.data.insert(name.clone(), AsylEnvEntry::Unevaluated(args[1].clone()));
+		}
 		Ok(args[0].clone())
 	}));
-	envi!(data, "fn", AsylExpr::ExtFn(|this, env, args| {
+	envi!(map, data, "fn", AsylExprValue::ExtFn(|this, env, args, _map| {
 		// todo: add a begin op and support more than 2 args here
 		// asal!(this, args, 2.., "at least two arguments");
+		// todo: create env here
 		asal!(this, args, 2..=2, "two arguments");
-		let args_exprs = ast!(this, args[0], List, "'list")?;
-		let lambda_args = vec![];
+		let args_exprs = ast!(this.clone(), args[0], List, "'list")?;
+		let mut lambda_args = vec![];
 		for arg in args_exprs {
-			lambda_args.push(ast!(this, arg, Symbol, "'symbol")?);
+			lambda_args.push(ast!(this.clone(), arg, Symbol, "'symbol")?.clone());
 		}
-		Ok(Spanned(AsylExpr::Lambda(AsylLambda {
-			args: Rc::new(lambda_args),
-			body: Rc::new(args[1].clone()),
-		}), this))
+		Ok(AsylExpr(AsylExprValue::Lambda(box AsylLambda {
+			args: lambda_args,
+			body: args[1].clone(),
+			env: env.clone(),
+		}), this, None))
 	}));
 	AsylEnv { data, parent: None }
 }
 
 #[derive(Debug)]
-pub enum AsylToken {
-	Symbol(String),
-	String(String),
+pub enum AsylTokenValue {
+	Symbol(MappedStr),
+	String(MappedStr),
 	Paren(bool),
 }
 
-#[derive(Debug, Clone)]
-pub struct Spanned<T>(pub T, pub AsylSpan);
+pub struct AsylToken(AsylTokenValue, AsylSpan);
+
+#[derive(Clone)]
+pub struct AsylExpr(pub AsylExprValue, pub AsylSpan, pub Option<Arc<RwLock<AsylEnv>>>);
 
 enum StartQuoteType {
 	Quote,
@@ -462,11 +480,11 @@ fn char_type(c: char) -> CharType {
 	}
 }
 
-pub fn tokenize(data: &str) -> AsylResult<Vec<Spanned<AsylToken>>> {
+pub fn tokenize(map: &mut StrMap, file_name: MappedStr, data: MappedStr) -> AsylResult<Vec<AsylToken>> {
 	let mut line = 1;
 	let mut col = 1;
-	let mut iter = data.chars().enumerate().map(|(i, v)| {
-		let res = (AsylPos(i, line, col), v);
+	let mut iter = data.get_ref().chars().enumerate().map(|(i, v)| {
+		let res = (AsylPos(i, line, col, file_name.clone(), data.clone()), v);
 		match v {
 			'\n' => { col = 1; line += 1; },
 			_ => { col += 1; }
@@ -480,7 +498,7 @@ pub fn tokenize(data: &str) -> AsylResult<Vec<Spanned<AsylToken>>> {
 			None => break,
 		};
 		match ch {
-			CharType::Paren(v) => out.push(Spanned(AsylToken::Paren(v), Some(AsylInSpan(chi, 1)))),
+			CharType::Paren(v) => out.push(AsylToken(AsylTokenValue::Paren(v), Some(AsylInSpan(chi, 1)))),
 			CharType::Comment => {
 				// discord until eol or eof
 				while match iter.next() {
@@ -497,7 +515,7 @@ pub fn tokenize(data: &str) -> AsylResult<Vec<Spanned<AsylToken>>> {
 				loop {
 					len += 1;
 					match iter.next() {
-						None => return Err(AsylError::UnexpectedEOF),
+						None => return Err(AsylError(AsylErrorType::UnexpectedEOF, None)),
 						Some((chi, ch)) => match ch {
 							// escapes
 							// shorthands: \n \t \r \a \e \0
@@ -506,7 +524,7 @@ pub fn tokenize(data: &str) -> AsylResult<Vec<Spanned<AsylToken>>> {
 							'\\' => {
 								len += 1;
 								match iter.next().and_then(|v| Some(v.1)) {
-									None => return Err(AsylError::UnexpectedEOF),
+									None => return Err(AsylError(AsylErrorType::UnexpectedEOF, None)),
 									Some('n') => res.push('\n'),
 									Some('t') => res.push('\t'),
 									Some('r') => res.push('\r'),
@@ -516,21 +534,21 @@ pub fn tokenize(data: &str) -> AsylResult<Vec<Spanned<AsylToken>>> {
 									Some('x') => {
 										let mut t = String::new();
 										len += 2;
-										t.push(iter.next().ok_or(AsylError::UnexpectedEOF)?.1);
-										t.push(iter.next().ok_or(AsylError::UnexpectedEOF)?.1);
+										t.push(iter.next().ok_or(AsylError(AsylErrorType::UnexpectedEOF, None))?.1);
+										t.push(iter.next().ok_or(AsylError(AsylErrorType::UnexpectedEOF, None))?.1);
 										let mut buf = [0u8; 1];
-										hex::decode_to_slice(&t, &mut buf).or_else(|_| Err(AsylError::InvalidEscape(Some(AsylInSpan(chi, 4)), format!("\\x{}", t))))?;
+										hex::decode_to_slice(&t, &mut buf).or_else(|_| Err(AsylError(AsylErrorType::InvalidEscape(format!("\\x{}", t)), Some(AsylInSpan(chi, 4)))))?;
 										res.push(char::from(buf[0]));
 									},
 									Some('{') => {
 										// unicode parsing :(
 										let mut esc = String::new();
 										loop { // 3rd level nested loop :)
-											if esc.len() > 8 { return Err(AsylError::InvalidEscape(Some(AsylInSpan(chi, 2 + esc.len())), format!("\\{{{}", esc))); }
+											if esc.len() > 8 { return Err(AsylError(AsylErrorType::InvalidEscape(format!("\\{{{}", esc)), Some(AsylInSpan(chi, 2 + esc.len())))); }
 											len += 1;
 											match iter.next() {
 												// 12 levels of indent oh no
-												None => return Err(AsylError::UnexpectedEOF),
+												None => return Err(AsylError(AsylErrorType::UnexpectedEOF, None)),
 												Some((_, '}')) => break,
 												Some((_, ch)) => esc.push(ch),
 											}
@@ -541,8 +559,8 @@ pub fn tokenize(data: &str) -> AsylResult<Vec<Spanned<AsylToken>>> {
 										}
 										// “`char` is always four bytes in size.”
 										let mut buf = [0u8; 4];
-										hex::decode_to_slice(&esc, &mut buf).or_else(|_| Err(AsylError::InvalidEscape(Some(AsylInSpan(chi, 2 + esc.len())), format!("\\{{{}", esc))))?;
-										res.push(char::from_u32(u32::from_be_bytes(buf)).ok_or_else(|| AsylError::InvalidEscape(Some(AsylInSpan(chi, 2 + esc.len())), format!("\\{{{}", esc)))?);
+										hex::decode_to_slice(&esc, &mut buf).or_else(|_| Err(AsylError(AsylErrorType::InvalidEscape(format!("\\{{{}", esc)), Some(AsylInSpan(chi.clone(), 2 + esc.len())))))?;
+										res.push(char::from_u32(u32::from_be_bytes(buf)).ok_or_else(|| AsylError(AsylErrorType::InvalidEscape(format!("\\{{{}", esc)), Some(AsylInSpan(chi.clone(), 2 + esc.len()))))?);
 									},
 									Some(tc) => res.push(tc),
 								}
@@ -557,7 +575,7 @@ pub fn tokenize(data: &str) -> AsylResult<Vec<Spanned<AsylToken>>> {
 						}
 					}
 				}
-				out.push(Spanned(AsylToken::String(res), Some(AsylInSpan(chi, len))));
+				out.push(AsylToken(AsylTokenValue::String(MappedStr::new(res, map)), Some(AsylInSpan(chi, len))));
 			},
 			CharType::Other(ch) => {
 				// consume until it's not other
@@ -575,7 +593,7 @@ pub fn tokenize(data: &str) -> AsylResult<Vec<Spanned<AsylToken>>> {
 					// we know this value so we discard it
 					iter.next();
 				}
-				out.push(Spanned(AsylToken::Symbol(res), Some(AsylInSpan(chi, len))));
+				out.push(AsylToken(AsylTokenValue::Symbol(MappedStr::new(res, map)), Some(AsylInSpan(chi, len))));
 			},
 			CharType::Whitespace => {},
 		}
@@ -586,7 +604,7 @@ pub fn tokenize(data: &str) -> AsylResult<Vec<Spanned<AsylToken>>> {
 fn merge_span(start: AsylSpan, end: AsylSpan) -> AsylSpan {
 	match start {
 		Some(a) => match end {
-			Some(b) => Some(AsylInSpan(a.0, b.1 + b.0.0 - a.0.0)),
+			Some(b) => Some(AsylInSpan(a.clone().0, b.1 + b.0.0 - a.0.0)),
 			None => Some(a),
 		},
 		None => match end {
@@ -596,7 +614,7 @@ fn merge_span(start: AsylSpan, end: AsylSpan) -> AsylSpan {
 	}
 }
 
-pub fn parse_all(tokens: &[Spanned<AsylToken>]) -> AsylResult<Vec<Spanned<AsylExpr>>> {
+pub fn parse_all(tokens: &[AsylToken]) -> AsylResult<Vec<AsylExpr>> {
 	let mut out = vec![];
 	let mut rest = tokens;
 	while rest.len() > 0 {
@@ -607,29 +625,29 @@ pub fn parse_all(tokens: &[Spanned<AsylToken>]) -> AsylResult<Vec<Spanned<AsylEx
 	Ok(out)
 }
 
-fn parse<'a>(tokens: &'a [Spanned<AsylToken>]) -> AsylResult<(Spanned<AsylExpr>, &'a [Spanned<AsylToken>])> {
-	let (token, rest) = tokens.split_first().ok_or(AsylError::UnexpectedEOF)?;
+fn parse<'a>(tokens: &'a [AsylToken]) -> AsylResult<(AsylExpr, &'a [AsylToken])> {
+	let (token, rest) = tokens.split_first().ok_or(AsylError(AsylErrorType::UnexpectedEOF, None))?;
 	match &token.0 {
-		AsylToken::Paren(true) => read_seq(token.1, rest),
-		AsylToken::Paren(false) => {
-			Err(AsylError::UnexpectedCloseParen(token.1))
+		AsylTokenValue::Paren(true) => read_seq(token.1.clone(), rest),
+		AsylTokenValue::Paren(false) => {
+			Err(AsylError(AsylErrorType::UnexpectedCloseParen, token.1.clone()))
 		},
-		AsylToken::String(data) => Ok((Spanned(AsylExpr::String(data.clone()), token.1), rest)),
-		AsylToken::Symbol(data) => if is_number(data.as_bytes()) {
-			Ok((parse_number(&data, token.1)?, rest))
+		AsylTokenValue::String(data) => Ok((AsylExpr(AsylExprValue::String(data.clone()), token.1.clone(), None), rest)),
+		AsylTokenValue::Symbol(data) => if is_number(data.get_ref().as_bytes()) {
+			Ok((parse_number(data.get_ref(), token.1.clone())?, rest))
 		} else {
-			Ok((Spanned(AsylExpr::Symbol(data.clone()), token.1), rest))
+			Ok((AsylExpr(AsylExprValue::Symbol(data.clone()), token.1.clone(), None), rest))
 		},
 	}
 }
 
-fn read_seq<'a>(start: AsylSpan, tokens: &'a [Spanned<AsylToken>]) -> AsylResult<(Spanned<AsylExpr>, &'a [Spanned<AsylToken>])> {
+fn read_seq<'a>(start: AsylSpan, tokens: &'a [AsylToken]) -> AsylResult<(AsylExpr, &'a [AsylToken])> {
 	let mut res = vec![];
 	let mut xs = tokens;
 	loop {
-		let (token, rest) = xs.split_first().ok_or(AsylError::UnexpectedEOF)?;
+		let (token, rest) = xs.split_first().ok_or(AsylError(AsylErrorType::UnexpectedEOF, None))?;
 		match token.0 {
-			AsylToken::Paren(false) => return Ok((Spanned(AsylExpr::List(res), merge_span(start, token.1)), rest)),
+			AsylTokenValue::Paren(false) => return Ok((AsylExpr(AsylExprValue::List(res), merge_span(start, token.1.clone()), None), rest)),
 			_ => {
 				let (exp, new_xs) = parse(&xs)?;
 				res.push(exp);
@@ -676,8 +694,8 @@ fn char_val(c: char, radix: u8) -> Option<u8> {
 	if val < radix {Some(val)} else {None}
 }
 
-fn parse_number(data: &str, span: AsylSpan) -> AsylResult<Spanned<AsylExpr>> {
-	let asyl_error = || AsylError::InvalidNumber(span, data.to_string());
+fn parse_number(data: &str, span: AsylSpan) -> AsylResult<AsylExpr> {
+	let asyl_error = || AsylError(AsylErrorType::InvalidNumber(data.to_string()), span.clone());
 	enum ParseState {
 		IntSign, IntBase, Int, Frac, ExpSign, Exp,
 	}
@@ -728,7 +746,7 @@ fn parse_number(data: &str, span: AsylSpan) -> AsylResult<Spanned<AsylExpr>> {
 				'e'|'E' if base == 10 => state = ParseState::ExpSign,
 				o => {
 					if int_len >= base_max {
-						return Err(AsylError::InvalidNumber(span, data.to_string()));
+						return Err(asyl_error());
 					}
 					int *= base as u64;
 					int |= char_val(o, base).ok_or_else(asyl_error)? as u64;
@@ -740,7 +758,7 @@ fn parse_number(data: &str, span: AsylSpan) -> AsylResult<Spanned<AsylExpr>> {
 				'e' => state = ParseState::ExpSign,
 				o => {
 					if frac_len >= 20 {
-						return Err(AsylError::InvalidNumber(span, data.to_string()));
+						return Err(asyl_error());
 					}
 					frac *= 10;
 					frac |= char_val(o, 10).ok_or_else(asyl_error)? as u64;
@@ -761,7 +779,7 @@ fn parse_number(data: &str, span: AsylSpan) -> AsylResult<Spanned<AsylExpr>> {
 				'_'|',' => {},
 				o => {
 					if exp_len >= 4 {
-						return Err(AsylError::InvalidNumber(span, data.to_string()));
+						return Err(asyl_error());
 					}
 					exp *= 10;
 					exp |= char_val(o, 10).ok_or_else(asyl_error)? as u64;
@@ -772,93 +790,122 @@ fn parse_number(data: &str, span: AsylSpan) -> AsylResult<Spanned<AsylExpr>> {
 	}
 	// and now, every sanity check in existance
 	match state {
-		ParseState::IntSign => Err(AsylError::InvalidNumber(span, data.to_string())),
-		ParseState::IntBase => Ok(Spanned(AsylExpr::Int(0), span)),
+		ParseState::IntSign => Err(asyl_error()),
+		ParseState::IntBase => Ok(AsylExpr(AsylExprValue::Int(0), span, None)),
 		ParseState::Int => if int_len > 0 {
 			if int_neg {
 				if int <= 0x1000000000000000u64 {
-					Ok(Spanned(AsylExpr::Int(-(int as i64)), span))
+					Ok(AsylExpr(AsylExprValue::Int(-(int as i64)), span, None))
 				} else {
-					Err(AsylError::InvalidNumber(span, data.to_string()))
+					Err(asyl_error())
 				}
 			} else if int < 0x1000000000000000u64 {
-				Ok(Spanned(AsylExpr::Int(int as i64), span))
+				Ok(AsylExpr(AsylExprValue::Int(int as i64), span, None))
 			} else {
-				Err(AsylError::InvalidNumber(span, data.to_string()))
+				Err(asyl_error())
 			}
 		} else {
-			Err(AsylError::InvalidNumber(span, data.to_string()))
+			Err(asyl_error())
 		},
 		ParseState::Frac => if int_len > 0{
-			Ok(Spanned(AsylExpr::Float((int as f64) + (frac as f64) / (10.0f64.powi(frac_len as i32))), span))
+			Ok(AsylExpr(AsylExprValue::Float((int as f64) + (frac as f64) / (10.0f64.powi(frac_len as i32))), span, None))
 		} else {
-			Err(AsylError::InvalidNumber(span, data.to_string()))
+			Err(asyl_error())
 		},
-		ParseState::ExpSign => Err(AsylError::InvalidNumber(span, data.to_string())),
+		ParseState::ExpSign => Err(asyl_error()),
 		ParseState::Exp => if int_len > 0 && exp_len > 0 {
 			let frac = (int as f64) + (frac as f64) / (10.0f64.powi(frac_len as i32));
-			Ok(Spanned(AsylExpr::Float(frac * 10.0f64.powi(exp as i32)), span))
+			Ok(AsylExpr(AsylExprValue::Float(frac * 10.0f64.powi(exp as i32)), span, None))
 		} else {
-			Err(AsylError::InvalidNumber(span, data.to_string()))
+			Err(asyl_error())
 		}
 	}
 }
 
-fn eval_args(args: &[Spanned<AsylExpr>], env: &mut AsylEnv) -> AsylResult<Vec<Spanned<AsylExpr>>> {
-	args.iter().map(|x| eval(x, env)).collect()
+fn eval_args(args: &[AsylExpr], env: &Arc<RwLock<AsylEnv>>, map: &mut StrMap) -> AsylResult<Vec<AsylExpr>> {
+	args.iter().map(|x| eval(x, env, map)).collect()
 }
 
-fn lookup_var(span: AsylSpan, name: &str, env: &mut AsylEnv) -> AsylResult<Spanned<AsylExpr>> {
-	match env.data.get(name) {
-		Some(v) => match v {
-			AsylEnvEntry::Unevaluated(exp) => {
-				let exp_eval = eval(&exp, env)?;
-				env.insert(name.clone(), AsylEnvEntry::Evaluated(exp_eval));
-				Ok(exp_eval)
+fn lookup_var(span: AsylSpan, name: &MappedStr, env: &Arc<RwLock<AsylEnv>>, map: &mut StrMap) -> AsylResult<AsylExpr> {
+	enum LookupResult {
+		Ok(AsylExpr),
+		Err(AsylError),
+		Recurse(Arc<RwLock<AsylEnv>>),
+		Insert(AsylExpr),
+	}
+	let res = {
+		let lock = env.read().map_err(|_| AsylError(AsylErrorType::Internal, span.clone()))?;
+		match lock.data.get(&name) {
+			Some(v) => match v {
+				AsylEnvEntry::Unevaluated(exp) => {
+					// lock.data.insert(name.to_string(), AsylEnvEntry::Evaluated(exp_eval.clone()));
+					LookupResult::Insert(exp.clone())
+				},
+				AsylEnvEntry::Evaluated(exp) => LookupResult::Ok(exp.clone()),
 			},
-			AsylEnvEntry::Evaluated(exp) => Ok(exp),
-		},
-		None => match env.parent {
-			Some(parent) => lookup_var(name, parent),
-			None => Err(AsylError::NotDefined(span, name.clone())),
+			None => match &lock.parent {
+				Some(parent) => LookupResult::Recurse(parent.clone()),
+				None => LookupResult::Err(AsylError(AsylErrorType::NotDefined(name.to_string()), span.clone())),
+			}
+		}
+	};
+	match res {
+		LookupResult::Ok(v) => Ok(v),
+		LookupResult::Err(v) => Err(v),
+		LookupResult::Recurse(parent) => lookup_var(span, name, &parent, map),
+		LookupResult::Insert(exp) => {
+			let value = eval(&exp, env, map)?;
+			let mut write_lock = env.write().map_err(|_| AsylError(AsylErrorType::Internal, span))?;
+			write_lock.data.insert(name.clone(), AsylEnvEntry::Evaluated(value.clone()));
+			Ok(value)
 		}
 	}
 }
 
 // this feels like the kind of thing that'd just desugar into like (lambda-exec lambda . args)
-fn lambda_env(span: AsylSpan, args: &[String], parent: &mut AsylEnv, arg_vals: &[AsylExpr]) -> AsylResult<AsylEnv> {
+fn lambda_env(span: AsylSpan, args: &[MappedStr], parent: &Arc<RwLock<AsylEnv>>, arg_vals: &[AsylExpr]) -> AsylResult<AsylEnv> {
 	if arg_vals.len() != args.len() {
-		Err(AsylError::ArgMismatch(span, format!("{} argument(s)", arg_vals.len())))
+		return Err(AsylError(AsylErrorType::ArgMismatch(format!("{} argument(s)", arg_vals.len())), span))
 	}
 	let mut data = HashMap::new();
 	for (k, v) in args.iter().zip(arg_vals.iter()) {
 		data.insert(k.clone(), AsylEnvEntry::Unevaluated(v.clone()));
 	}
-	Ok(AsylEnv { data, parent })
+	Ok(AsylEnv { data, parent: Some(parent.clone()) })
 }
 
 // todo: tail recursion optimization
-pub fn eval(exp: &Spanned<AsylExpr>, env: &mut AsylEnv) -> AsylResult<Spanned<AsylExpr>> {
+pub fn eval(exp: &AsylExpr, env: &Arc<RwLock<AsylEnv>>, map: &mut StrMap) -> AsylResult<AsylExpr> {
+	let env = match &exp.2 {
+		Some(v) => v,
+		None => env,
+	};
 	match &exp.0 {
-		AsylExpr::Symbol(v) => lookup_var(exp.1, v, env),
-		AsylExpr::List(list) => {
-			let first = list.first().ok_or(AsylError::InvalidCall(exp.1))?;
+		AsylExprValue::Symbol(v) => lookup_var(exp.1.clone(), v, env, map),
+		AsylExprValue::List(list) => {
+			let first = list.first().ok_or(AsylError(AsylErrorType::InvalidCall, exp.1.clone()))?;
 			let args = &list[1..];
-			let res = eval(first, env)?;
+			let res = eval(first, env, map)?;
 			match res.0 {
-				AsylExpr::ExtFn(f) => {
-					Ok(f(exp.1, env, args)?)
+				AsylExprValue::ExtFn(f) => {
+					Ok(f(exp.1.clone(), env, args, map)?)
 				},
-				AsylExpr::Lambda(f) => {
-					let new_env = lambda_env(exp.1, f.args, env, args)?;
-					eval(&f.body, &mut new_env)
+				AsylExprValue::Lambda(f) => {
+					let new_env = Arc::new(RwLock::new(
+						lambda_env(exp.1.clone(), &f.args, &f.env, &args
+							.iter()
+							.map(|AsylExpr(a, b, _)| AsylExpr(a.clone(), b.clone(), Some(env.clone())))
+							.collect::<Vec<_>>()
+						)?
+					));
+					eval(&f.body, &new_env, map)
 				},
 				// shorthand for an identity / force function
 				// for instance (define var 12) returns var
 				// but ((define var 12)) returns 12
-				other => eval(other, env),
+				_ => eval(exp, env, map),
 			}
 		},
-		other => Ok(Spanned(other.clone(), exp.1))
+		other => Ok(AsylExpr(other.clone(), exp.1.clone(), exp.2.as_ref().map(|v| v.clone())))
 	}
 }