+++ changes.md
@@ -1,7 +1,5 @@
 # some changes i need to make but don't feel like doing right now
 
-- make env's Rc'd
-- make symbols store references to their env
-- create lambda env's at definition rather than at call (in `fn` rather than eval)
 - replace error strings with appropriate types (Range, AsylType, etc.)
 - fix all the errors :(
+- forward panics?